Claude- Google Drive MCP/requirements.txt
mcp>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
anthropic>=0.40.0
requests>=2.31.0
